payment and order bugs

+++ src/main/java/com/thelastcodebenders/follower/service/BankAccountService.java
@@ -2,6 +2,7 @@ package com.thelastcodebenders.follower.service;
 
 import com.thelastcodebenders.follower.model.BankAccount;
 import com.thelastcodebenders.follower.repository.BankAccountRepository;
+import com.thelastcodebenders.follower.repository.PaymentNotificationRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -14,9 +15,12 @@ public class BankAccountService {
     private static final Logger log = LoggerFactory.getLogger(BankAccountService.class);
 
     private BankAccountRepository bankAccountRepository;
+    private PaymentNotificationRepository paymentNotificationRepository;
 
-    public BankAccountService(BankAccountRepository bankAccountRepository){
+    public BankAccountService(BankAccountRepository bankAccountRepository,
+                              PaymentNotificationRepository paymentNotificationRepository){
         this.bankAccountRepository = bankAccountRepository;
+        this.paymentNotificationRepository = paymentNotificationRepository;
     }
 
     public void save(BankAccount bankAccount){
@@ -34,9 +38,20 @@ public class BankAccountService {
 
     public boolean deleteAccount(long id){
         try {
+            BankAccount bankAccount = findById(id);
+
+            if (bankAccount == null)
+                return false;
+
+            if (paymentNotificationRepository.countByBankAccount(bankAccount)>0){
+                throw new RuntimeException("Bu banka hesabına bağlı olarak oluşturulmuş ödeme bildirimleri mevcut, bu yüzden silinemez !");
+            }
+
             bankAccountRepository.deleteById(id);
             return true;
         }catch (Exception e){
+            if (e instanceof RuntimeException)
+                throw e;
             log.error("Delete account exception ! - " + e.getMessage());
             return false;
         }
